requirements.txt
APScheduler==3.7.0
Pillow
aiohttp
asyncio
bs4
dnspython
emoji
gitpython
google-search==1.1.1
google==3.0.0
googlesearch-python== 1.0.1
gpytranslate
gtts
heroku3
humanize
lottie
motor
numpy
prettytable
psutil
psycopg2-binary
ptable
pymediainfo
pyrogram
pytgcalls==3.0.0.dev22
python-dotenv
requests
speedtest-cli==2.1.3
sqlalchemy==1.3.23
telegraph
textblob
tgcrypto
tracemoepy
opencv-python-headless
urllib3
uvloop
wget==3.2
youtube-search-python
yt-dlp
httpx[http2]
openai
